third_party/blink/web_tests/svg/filters/feConvolveMatrix-order-error-reporting-expected.txt

CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("0")
CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("0.49999999")
CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("0.5")
CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: A negative value is not valid. ("-1")
CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("1 0")
CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("1 0.49999999")
CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: A value of zero is not valid. ("1 0.5")
CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: A negative value is not valid. ("1 -1")
CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: Expected integer, "1 one".
CONSOLE ERROR: Error: <feConvolveMatrix> attribute order: Expected integer, "one two".
